chore: rename ReconnectClient to Client, Client to RawClient
