+++ socketio/src/client/callback.rs
@@ -3,12 +3,12 @@ use std::{
     ops::{Deref, DerefMut},
 };
 
-use super::Client;
+use super::RawClient;
 use crate::{Event, Payload};
 
-pub(crate) type SocketCallback = Box<dyn for<'a> FnMut(Payload, Client) + 'static + Sync + Send>;
+pub(crate) type SocketCallback = Box<dyn for<'a> FnMut(Payload, RawClient) + 'static + Sync + Send>;
 pub(crate) type SocketAnyCallback =
-    Box<dyn for<'a> FnMut(Event, Payload, Client) + 'static + Sync + Send>;
+    Box<dyn for<'a> FnMut(Event, Payload, RawClient) + 'static + Sync + Send>;
 
 pub(crate) struct Callback<T> {
     inner: T,
@@ -23,7 +23,7 @@ impl Debug for Callback<SocketCallback> {
 }
 
 impl Deref for Callback<SocketCallback> {
-    type Target = dyn for<'a> FnMut(Payload, Client) + 'static + Sync + Send;
+    type Target = dyn for<'a> FnMut(Payload, RawClient) + 'static + Sync + Send;
 
     fn deref(&self) -> &Self::Target {
         self.inner.as_ref()
@@ -39,7 +39,7 @@ impl DerefMut for Callback<SocketCallback> {
 impl Callback<SocketCallback> {
     pub(crate) fn new<T>(callback: T) -> Self
     where
-        T: for<'a> FnMut(Payload, Client) + 'static + Sync + Send,
+        T: for<'a> FnMut(Payload, RawClient) + 'static + Sync + Send,
     {
         Callback {
             inner: Box::new(callback),
@@ -56,7 +56,7 @@ impl Debug for Callback<SocketAnyCallback> {
 }
 
 impl Deref for Callback<SocketAnyCallback> {
-    type Target = dyn for<'a> FnMut(Event, Payload, Client) + 'static + Sync + Send;
+    type Target = dyn for<'a> FnMut(Event, Payload, RawClient) + 'static + Sync + Send;
 
     fn deref(&self) -> &Self::Target {
         self.inner.as_ref()
@@ -72,7 +72,7 @@ impl DerefMut for Callback<SocketAnyCallback> {
 impl Callback<SocketAnyCallback> {
     pub(crate) fn new<T>(callback: T) -> Self
     where
-        T: for<'a> FnMut(Event, Payload, Client) + 'static + Sync + Send,
+        T: for<'a> FnMut(Event, Payload, RawClient) + 'static + Sync + Send,
     {
         Callback {
             inner: Box::new(callback),
