+++ socketio/src/client/mod.rs
@@ -1,10 +1,11 @@
 mod builder;
-mod client;
+mod raw_client;
 
 pub use builder::ClientBuilder;
 pub use builder::TransportType;
 pub use client::Client;
+pub use raw_client::RawClient;
 
 /// Internal callback type
 mod callback;
-mod reconnect;
+mod client;
